+++ math/src/arithmetic.c
@@ -1,14 +1,14 @@
 /**
  * @author Ytalo Ramon
  * @date 13/06/2021
- * 
-*/
+ *
+ */
 
 #include "../include/arithmetic.h"
 
-int divmodular(int a, int b){
-    
+size_t divmodular(int a, int b) {
+
     int v = a % b;
-    
+
     return v >= 0 ? v : v + b;
 }
